requirements.txt
requests==2.31.0
spotipy==2.24.0
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1
flask==2.3.3
gunicorn==21.2.0 